ISSUE #4827 - FIxed limits when zooming
