+++ frontend/package.json
@@ -51,7 +51,7 @@
 		"@mui/system": "5.4.1",
 		"@mui/x-date-pickers": "5.0.3",
 		"@popperjs/core": "2.6.0",
-		"@svgedit/svgcanvas": "^7.2.2",
+		"@svgedit/svgcanvas": "7.2.2",
 		"@webscopeio/react-textarea-autocomplete": "4.9.0",
 		"axios": "1.6.5",
 		"byte-size": "8.1.0",
@@ -77,7 +77,7 @@
 		"normalize.css": "8.0.1",
 		"normalizr": "3.4.0",
 		"pdf-lib": "1.17.1",
-		"panzoom": "^9.4.3",
+		"panzoom": "9.4.3",
 		"prop-types": "15.5.8",
 		"query-string": "6.2.0",
 		"react": "17.0.2",
